requirements.txt
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
docutils==0.19
idna==3.4
importlib-metadata==6.5.0
jaraco.classes==3.2.3
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
pip==23.0.1
pkginfo==1.9.6
Pygments==2.15.1
python-dotenv==1.0.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.4
setuptools==67.6.1
six==1.16.0
tqdm==4.65.0
twine==4.0.2
urllib3==1.26.15
webencodings==0.5.1
wheel==0.40.0
zipp==3.15.0
